LOS PN Fixed
Fixed the algorithm for LOS Proportional Navigation. The drone now moves in an arc towards the obstacle depending on the starting position. The Proportional Gain constant will need to be tuned in order to ensure that this arc is sufficiently flat. 

TODO: A feedback system for when the drone has passed through an obstacle needs to be implemented