LOS PN Demo video
Added a demo video of the drone navigating to the obstacle (a red cube) using Proportional Navigation in the Block environment. This demo assumes the coordinates of the obstacle and the drone are known, while these will need to be extrapolated from GPS data and camera images in the full application.