Image depth estimation calibration
Added image_calibration.py which fits an exponential function to the ratios of the distance to and size of an obstacle in meters and in 'pixels'.  This allows us to estimate the distance to an obstacl;e using the fitted function and the known size of the obstacle in meters and pixels (calculated from the height of the bounding box). When a bounding box is not square, the higher of its width or height should be used. The accuracy of this function is poor past about 5 meters and when the drone is not moving directly through the obstacle.